restructure fingerprint
