+++ catlearn/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.5.0.dev1"
+__version__ = "0.5.0.dev2"
