+++ catlearn/featurize/base.py
