+++ catlearn/featurize/neighbor_matrix.py
