+++ catlearn/fingerprint/adsorbate.py
@@ -1,15 +1,18 @@
 """Slab adsorbate fingerprint functions for machine learning."""
 import numpy as np
+import collections
+
 from ase.atoms import string2symbols
 from ase.data import ground_state_magnetic_moments as gs_magmom
 from ase.data import atomic_numbers, chemical_symbols
-from .periodic_table_data import (get_mendeleev_params, n_outer,
-                                  list_mendeleev_params,
-                                  default_params, get_radius,
-                                  electronegativities,
-                                  block2number, make_labels)
-import collections
-from .base import BaseGenerator, check_labels
+
+from catlearn.featurize.periodic_table_data import (get_mendeleev_params,
+                                                    n_outer,
+                                                    list_mendeleev_params,
+                                                    default_params, get_radius,
+                                                    electronegativities,
+                                                    block2number, make_labels)
+from catlearn.featurize.base import BaseGenerator, check_labels
 
 
 default_adsorbate_fingerprinters = ['mean_chemisorbed_atoms',
